fix bugs in new category identification and existing feed item checks
* Code that determines what categories should be added to the database
was inconsistent, with some parts working with Category objects and
others working with the string value of the category.

* Date being used for matching existing articles was the RSS
timestamp rather than the SQL datetime. Moved the conversion
from RSS to SQL format to before the query.

* Moved feed_retrieval_id computation to earlier in code because
the feed_item_counts decrement that happens when a feed item
already exists needs it.

* Removed excessive spaces between a dot and a method name.