fix object formatting bug, fix BS4 warning, add logging, revamp scripts
This commit contains the folloing changes:

* Fixed off-by-one references to `format()` arguments in formatting
`FeedRetrieval` objects.

* Fixed warning being emitted by BeautifulSoup because no parser
was being specified. Opted for `html.parser`.

* Added INFO level logging statements throughout the code. Set up
scripts to log with timestamp with a file by the name of newscorpus.log.

* Renamed scirpts `setup_db.py` and `process_feeds.py` to `nc_setup_db` and
`nc_process_feeds` respectively.

* Added positional command line arguments to both `nc_setup_db` and
`nc_process_feeds` so that the logging direcotry and database
configuration can be specified on the command line.

* Bumped package revision to 0.1.0 because of name changes to scripts.

* Updated documentation in `README.md` to reflect the changes.