fix timezone correction for real
The previous addition of code using pytz revealed a misunderstanding
on my part. From the documentation for this package:

> Unfortunately using the tzinfo argument of the standard datetime
constructors ‘’does not work’’ with pytz for many timezones.

The proper approach is to use the localize() method on a pytz-generated
timezone, feeding it a naive Python datetime object. Then you can
call the astimezone() method on the datetime object with an instance
of a UTC timezone.

Also realized that the Etc/GMT offsets work opposite of what you
would expect. That is, Etc/GMT-4 => four hours ahead of GMT.
Had to reverse the polarity on a bunch of entries in the timezone
map.